Cost calculator almost done
Created div and added logic in event listener to loop through and update cost after each click. BUT, Main Event isn't calculated correctly. I think it's because it doesn't have a  data-day-and-time attribute. Since this is part of the logic in the if statement, it doesn't work. Do I need separate if statement? That's not very DRY.

Also, why am I dynamically creating the total cost HTML? Seems kludgy.